Code clean up and restructure, not sure on messages folder
