+++ foundationz_backend/src/modules/graphql/auth/auth.ts
@@ -1,4 +1,4 @@
-import { extendType, nonNull, objectType, stringArg } from "nexus";
+import { extendType, nonNull, stringArg } from "nexus";
 import { User } from "@orm/user";
 import { validatePassword } from "@auth/crypto";
 import { AuthenticationError } from "apollo-server-express";
@@ -38,21 +38,6 @@ export const UserLogin = extendType({
   },
 });
 
-export const UserAuthType = objectType({
-  name: "UserAuth",
-  definition(t) {
-    t.nonNull.string("id");
-    t.nonNull.string("email");
-  },
-});
-
-export const StandardMessage = objectType({
-  name: "Message",
-  definition(t) {
-    t.nonNull.string("message");
-  },
-});
-
 export const UserLogout = extendType({
   type: "Query",
   definition(t) {
@@ -72,27 +57,6 @@ export const UserLogout = extendType({
             message: "Logged out",
           },
         ];
-      },
-    });
-  },
-});
-
-export const WhoAmI = extendType({
-  type: "Query",
-  definition(t) {
-    t.nonNull.list.field("whoAmI", {
-      type: "User",
-
-      resolve: async (_, {}, context) => {
-        if (!context.session.email) {
-          throw new AuthenticationError(`You are not signed in!`);
-        }
-        const user = await User.query().where("id", context.session.userid);
-        if (!user) {
-          throw new AuthenticationError(`You are not signed in!`);
-        }
-        console.log(context.user);
-        return user || [];
       },
     });
   },
