+++ foundationz_backend/src/modules/graphql/index.ts
@@ -6,3 +6,5 @@ export * from "./resource/ResourceType";
 export * from "./project/ProjectType";
 export * from "./product/ProductType";
 export * from "./auth/auth";
+export * from "./auth/WhoAmI";
+export * from "./messages/MessageTypes";
