+++ foundationz_backend/src/modules/graphql/messages/MessageTypes.ts
@@ -1,0 +1,8 @@
+import { objectType } from "nexus";
+
+export const StandardMessage = objectType({
+  name: "Message",
+  definition(t) {
+    t.nonNull.string("message");
+  },
+});
