+++ foundationz_backend/src/modules/graphql/user/UserType.ts
@@ -29,3 +29,11 @@ export const UserType = objectType({
     });
   },
 });
+
+export const UserAuthType = objectType({
+  name: "UserAuth",
+  definition(t) {
+    t.nonNull.string("id");
+    t.nonNull.string("email");
+  },
+});
